--- apps/nextjs/src/components/features/tasks/index.ts
@@ -1,8 +1,0 @@
-export * from './tasks-board';
-export * from './tasks-calendar';
-export * from './tasks-container';
-export * from './tasks-files';
-export * from './tasks-header';
-export * from './tasks-list';
-export * from './tasks-modals';
-export * from './tasks-provider';
