+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-item/tasks-list-subtask-list/skeleton-list.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import { memo } from 'react';
-import { TasksListRow } from '@/components/features/tasks';
+import { TasksListRow } from '@/components/features/tasks/tasks-list/tasks-list-row/tasks-list-row';
 import { Flex } from '@/components/ui/flex';
 import { Skeleton } from '@/components/ui/skeleton';
