+++ apps/nextjs/src/components/pages/inbox/components/activity/activity.tsx
@@ -13,7 +13,7 @@ import { InboxListContent } from '@/components/features/inbox/inbox-list-content
 import { InboxRight } from '@/components/features/inbox/inbox-right';
 import { InboxSkeleton } from '@/components/features/inbox/inbox-skeleton';
 import { TaskDetailSide } from '@/components/features/task-details/task-detail-side';
-import { TasksProvider } from '@/components/features/tasks';
+import { TasksProvider } from '@/components/features/tasks/tasks-provider/tasks-provider';
 import { Flex } from '@/components/ui/flex';
 import { useInboxActivityPageQuery } from '@/hooks/queries/app';
 import { getInboxDetailId, isInboxDetailURL } from '@/router';
