+++ apps/nextjs/src/components/pages/projects/components/board/skeleton-board/skeleton-board-header.tsx
@@ -1,5 +1,8 @@
 import { memo } from 'react';
-import { TasksHeader, TasksHeaderRight } from '@/components/features/tasks';
+import {
+  TasksHeader,
+  TasksHeaderRight,
+} from '@/components/features/tasks/tasks-header';
 import type { FlexProps } from '@/components/ui/flex';
 import { Skeleton } from '@/components/ui/skeleton';
