+++ apps/nextjs/src/components/pages/projects/components/list/list.tsx
@@ -1,23 +1,23 @@
 import { memo, useCallback } from 'react';
 import { TaskDetailDrawer } from '@/components/features/task-details/task-detail-drawer';
+import { TasksContainer } from '@/components/features/tasks/tasks-container/tasks-container';
 import {
   AddTaskButton,
   CustomizeButton,
   CustomizeMenu,
   IncompleteTasksMenu,
   MoreActionMenu,
-  TasksContainer,
   TasksHeader,
   TasksHeaderLeft,
   TasksHeaderRight,
-  TasksList,
-  TasksListBody,
-  TasksListContent,
-  TasksListHeader,
-  TasksListHorizontalScrollBorder,
-  TasksListLayout,
-  useTasksListDetail,
-} from '@/components/features/tasks';
+} from '@/components/features/tasks/tasks-header';
+import { TasksList } from '@/components/features/tasks/tasks-list';
+import { TasksListBody } from '@/components/features/tasks/tasks-list/tasks-list-body/tasks-list-body';
+import { TasksListContent } from '@/components/features/tasks/tasks-list/tasks-list-content/tasks-list-content';
+import { TasksListHeader } from '@/components/features/tasks/tasks-list/tasks-list-header/tasks-list-header';
+import { TasksListHorizontalScrollBorder } from '@/components/features/tasks/tasks-list/tasks-list-horizontal-scroll-border/tasks-list-horizontal-scroll-border';
+import { TasksListLayout } from '@/components/features/tasks/tasks-list/tasks-list-layout/tasks-list-layout';
+import { useTasksListDetail } from '@/components/features/tasks/tasks-list/use-tasks-list-detail';
 import { Flex } from '@/components/ui/flex';
 import { getProjectsDetailId, isProjectsDetailURL, useRouter } from '@/router';
 import { useProjectsProjectId } from '@/store/app/projects/project';
