+++ apps/nextjs/src/components/pages/projects/components/overview/skeleton-overview.tsx
@@ -1,5 +1,8 @@
 import { memo } from 'react';
-import { TasksHeader, TasksHeaderRight } from '@/components/features/tasks';
+import {
+  TasksHeader,
+  TasksHeaderRight,
+} from '@/components/features/tasks/tasks-header';
 import { Flex, type FlexProps } from '@/components/ui/flex';
 import { Skeleton } from '@/components/ui/skeleton';
 import { Stack } from '@/components/ui/stack';
